requirements.txt
omegaconf
Pillow
einops
transformers
trimesh
rembg
rtree
huggingface-hub
imageio[ffmpeg]
gradio
xatlas
moderngl
three
numpy
torch
torchvision
pandas
matplotlib
uvicorn
fastapi
onnxruntime
onnxruntime-gpu
scikit-image
scipy
scikit-learn
tqdm
nbtlib
apscheduler
cuda-python
CMake
git+https://github.com/tatsy/torchmcubes.git